refacto(FemModule.cc): Every bilinearAssembly method is now in a file depending of the datastructure it is used for
